(refactor) Address PR #27 review comments
- Restore indentation whitespace on blank lines in DemoConditionGenerator
  to match house style (sibling Randomizer.java pattern).
- Extract DrugOrderSpec inner class out of DemoOrderGenerator into a
  top-level DrugOrderDescriptor file matching the existing *Descriptor
  naming convention.
- Drop bumpH2IdentityCounters and the three seed* helpers from
  ReferenceDemoDataActivatorTest; move that reference data into a sibling
  DBUnit dataset (devanModiSeedTestDataset.xml) loaded via executeDataSet.
  PKs >= 100000 in the new dataset remove the need for the H2 identity
  bump entirely.
